Fix for OWF-6588 - Can not remove the description of a dashboard even though it is not a required field.
I fixed this by programmatically changing an empty description field to one with a blank space (which was already our work around) when the user clicks save.  

Under the hood it was sending a null value over to the code that saves the value to the database when the user clears the description field, and I did not get into it deep enough to see exactly where it is reverting back to the old value when it sees a null value for description.  There was a reference to the original value in a "modified" field at a pretty low level, so I was assuming that at some point it saw the null value and reverted back to that modified value.