Fix for OWF-6111- Dashboards modal window allows focus to leave window through keyboard.
The "circular tabbing" was not being setup properly the 2nd time the Dashboard Switcher was populated, because the "setupFocus" function in the base class of CircularFocus was never being triggered by the derived DashboardSwitcher after the initial time it is loaded.  This caused the tabbing index to go off the modal dialog and stay in the background.

I fixed this by adding in the same call that happens on the initial load to happen on the viewready event, which happens every time the dialog is populated.