Fix for OWF-6579 - Add menu configuration text box runs beyond widgets boundaries.
I explicitly set the textArea height to be the height of the msgBox - 100 pixels after it is being initialized to be too large.  The textArea height was being set to the widget's height - 150 pixels.  Since tabbed and fit pane dashboards can have full screen widgets, the text area was very large.

There was already a fix in there to fix the width which used to have the same problem, and now this fixes the height as well.