fix error: java.util.zip.ZipException: error in opening zip file
AntClassLoader is going through each of the files in the classpath, the
AntClassLoader.getResourceURL() method is called for each of them, and
if the resource is a valid file, it attempts to open it as a ZipFile.
This is a problem as not all files in compass-1.0/lib directory are zip
files. Simply renaming it to libs and updating .groovy files in the
plugin fixes this problem.