OP-63: Added in the new button graphics and the graphics changes based on the user interaction.  I also fixed 2 minor bugs in the designer.
The bugs were when the white color would show around the buttons after a failed drop.  The other one was when you would reset a design after it had a split pane and a pane type, and when you split it from a clear slate, it would have the pane type in it.  It now has no pane type in it after a reset and split.