OP-1589: The word "value" appears to be some kind of keyword in the dbm plugin so I changed it to "VALUE".
So when liquibase sees the table column name it added the column to the prodDB.script file as "value" with the quotes around it, and it was being case sensitive when trying to call it by "VALUE" it works.  I changed the changelog to use the table name "VALUE" instead of "value" and it works locally now instead of throwing the same error as jenkins when trying to start up the server, so hopefully this fixes Jenkins as well.