Fix issue with removing automatic groups
There was a bug where automatic groups would not be removed from a user
if they should now have zero.  This bug was introduced by a botched fix
for an old bug where an invalid SQL query would be constructed in the
same scenario.  This commit undoes the original fix and adds a better
one so that neither issue occurs